+++ project_rates/urls.py
@@ -1,5 +1,8 @@
 from django.urls import path
 
-from project_rates.views import RateProject
+from project_rates.views import RateProject, RateProjects
 
-urlpatterns = [path("rate/<int:project_id>", RateProject.as_view())]
+urlpatterns = [
+    path("rate/", RateProject.as_view()),
+    path("<int:program_id>", RateProjects.as_view()),
+]
